requirements.txt
#google-genai>=1.0.0
groq>=0.9.0
ollama>=0.1.0
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
pandas>=1.5.0
scikit-learn>=1.0.0
numpy>=1.23.0
python-dotenv>=0.20.0
yfinance>=0.2.0
ddgs>=0.1.0
openpyxl>=3.0.0